+++ stories/illustrations.js
@@ -21,6 +21,24 @@ const itemStyles = {
 };
 
 storiesOf('Illustrations', module)
+  .add('36x24', () => (
+    <Box style={gridStyles}>
+      {Object.keys(Illustrations).map((key, index) => {
+        if (key.includes('36X24')) {
+          const IllustrationToRender = Illustrations[key];
+
+          return (
+            <Box key={index} style={itemStyles} {...itemProps}>
+              <Box flex="0 0 36px">
+                <IllustrationToRender />
+              </Box>
+              <TextSmall marginLeft={3}>{key}</TextSmall>
+            </Box>
+          );
+        }
+      })}
+    </Box>
+  ))
   .add('48x48', () => (
     <Box style={gridStyles}>
       {Object.keys(Illustrations).map((key, index) => {
@@ -30,6 +48,24 @@ storiesOf('Illustrations', module)
           return (
             <Box key={index} style={itemStyles} {...itemProps}>
               <Box flex="0 0 48px">
+                <IllustrationToRender />
+              </Box>
+              <TextSmall marginLeft={3}>{key}</TextSmall>
+            </Box>
+          );
+        }
+      })}
+    </Box>
+  ))
+  .add('60x60', () => (
+    <Box style={gridStyles}>
+      {Object.keys(Illustrations).map((key, index) => {
+        if (key.includes('60X60')) {
+          const IllustrationToRender = Illustrations[key];
+
+          return (
+            <Box key={index} style={itemStyles} {...itemProps}>
+              <Box flex="0 0 60px">
                 <IllustrationToRender />
               </Box>
               <TextSmall marginLeft={3}>{key}</TextSmall>
@@ -66,6 +102,24 @@ storiesOf('Illustrations', module)
           return (
             <Box key={index} style={itemStyles} {...itemProps}>
               <Box flex="0 0 120px">
+                <IllustrationToRender />
+              </Box>
+              <TextSmall marginLeft={3}>{key}</TextSmall>
+            </Box>
+          );
+        }
+      })}
+    </Box>
+  ))
+  .add('240x240', () => (
+    <Box style={gridStyles}>
+      {Object.keys(Illustrations).map((key, index) => {
+        if (key.includes('240X240')) {
+          const IllustrationToRender = Illustrations[key];
+
+          return (
+            <Box key={index} style={itemStyles} {...itemProps}>
+              <Box flex="0 0 240px">
                 <IllustrationToRender />
               </Box>
               <TextSmall marginLeft={3}>{key}</TextSmall>