+++ CHANGELOG.md
@@ -4,13 +4,27 @@
 
 ### Changed
 
-### Changed
-
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+### Dependency updates
+
+## [0.28.1] - 2019-09-02
+
+### Added
+
+- `Box`: added `textAlign` prop. Possible values are `center`, `left` & `right`. ([@driesd](https://github.com/driesd) in [#673](https://github.com/teamleadercrm/ui/pull/673))
+
+### Changed
+
+- `Select`: changed font smoothing css props from kebab-case to camelCase to fix emotion warnings. ([@driesd](https://github.com/driesd) in [#672](https://github.com/teamleadercrm/ui/pull/672))
+
+### Dependency updates
+
+- `@teamleader/ui-illustrations` from `^0.0.20` to `^0.0.21`. ([@driesd](https://github.com/driesd) in [#674](https://github.com/teamleadercrm/ui/pull/674))
 
 ## [0.28.0] - 2019-08-29